chore: overwrite `focus-test` with `new-focus-test` now that it's working
